web-purchase-bot/requirements.txt
playwright==1.55.0
PyYAML==6.0.2
rich==13.9.2
